do not print full circuit in metadata
