+++ povm_toolbox/library/metadata/povm_metadata.py
@@ -12,8 +12,10 @@
 
 from __future__ import annotations
 
+import dataclasses
 from dataclasses import dataclass
 
+import numpy as np
 from qiskit.circuit import QuantumCircuit
 
 from ..povm_implementation import POVMImplementation
@@ -33,3 +35,24 @@ class POVMMetadata:
     measurement circuit generated by the POVM implementation used. This is the quantum
     circuit that is eventually sent to the internal :class:`.BaseSamplerV2`.
     """
+
+    def __repr__(self):
+        """Implement the default ``__repr__`` method to avoid printing large objects.
+
+        E.g., the attribute ``composed_circuit`` is a quantum circuit. With the default
+        ``dataclass.__repr__``, it would be entirely drawn. As this is recursive,
+        the full circuit would be printed when printing the :class:`.PrimitiveResult`
+        object returned by the :meth:`.POVMSampler.run` method. The redefinition
+        here avoids this.
+        """
+        lst_fields = []
+        for field in dataclasses.fields(self):
+            f_name = field.name
+            f_val = getattr(self, field.name)
+            if isinstance(f_val, np.ndarray):
+                f_val = f'np.ndarray<{",".join(map(str, f_val.shape))}>'
+            elif isinstance(f_val, QuantumCircuit):
+                f_val = f_val.__repr__()
+            lst_fields.append((f_name, f_val))
+        f_repr = ", ".join(f"{name}={value}" for name, value in lst_fields)
+        return f"{self.__class__.__name__}({f_repr})"
